// frontend/src/components/dashboard/poster-generator/types.ts
export interface PosterData {
    title: string;
    price2Person: string;
    price3Person: string;
    price4Person: string;
    date: string;
    guideName: string;
    guidePhone: string;
    isIdentityVerified: boolean;
    backgroundImage: string;
    frameStyle: string;
    fontStyle: string;
    hotelMecca: string;
    hotelMedina: string;
    guideImage?: string;
    urgencyText?: string;
}